Code for SR validation.
For validating the sensitivity reconstruction idea, I am working on a 2D heat conduncion problem. The boundary conditions are fixed at 0 with a hot surface inside defined by immerse boundary method. This code is now capabale of finding the temperature inside the domain. Next step is working on the SA.